fonction populate

+++ controller/post.controller.js
@@ -1,7 +1,8 @@
 /* 
 Imports
 */
-    const Models = require('../models/index')
+    const Models = require('../models/index');
+    const { cryptData, decryptData } = require('../services/crypto.service');
 //
 
 /* 
@@ -10,7 +11,7 @@ Functions
     // CRUD: create one
     const createOne = req => {
         return new Promise( (resolve, reject) => {
-            // Get all post from MongoDB
+            // Create new object
             Models.post.create(req.body)
             .then( data => resolve(data) )
             .catch( err => reject(err) )
@@ -20,12 +21,23 @@ Functions
     // CRUD: read all posts
     const readAll = () => {
         return new Promise( (resolve, reject) => {
-            // Get all post from MongoDB
-            Models.post.find( (err, data) => {
-                // Check err
-                return err
-                ? reject(err)
-                : resolve(data)
+            // Make populated request
+            Models.post.find()
+            .populate({ 
+                path: 'author',
+                select: ['firstname', 'lastname', 'email']
+            })
+            .populate({ path: 'comment' })
+            .exec( (err, data) => {
+                // Check error
+                if( err ){ return reject(err) }
+                else{
+                    // Decrypt user data
+                    decryptData(data.author, 'firstname', 'lastname')
+
+                    // Send back data
+                    return resolve(data)
+                }
             })
         })
     }
@@ -33,12 +45,23 @@ Functions
     // CRUD: read one
     const readOne = req => {
         return new Promise( (resolve, reject) => {
-            // Get all post from MongoDB
-            Models.post.findById(req.params.id, (err, data) => {
-                // Check err
-                return err
-                ? reject(err)
-                : resolve(data)
+            // Make populated request
+            Models.post.findById(req.params.id)
+            .populate({ 
+                path: 'author',
+                select: ['firstname', 'lastname', 'email']
+            })
+            .populate({ path: 'comment' })
+            .exec( (err, data) => {
+                // Check error
+                if( err ){ return reject(err) }
+                else{
+                    // Decrypt user data
+                    decryptData(data.author, 'firstname', 'lastname')
+
+                    // Send back data
+                    return resolve(data)
+                }
             })
         })
     }
@@ -46,12 +69,23 @@ Functions
     // CRUD: update one
     const updateOne = req => {
         return new Promise( (resolve, reject) => {
-            // Get all post from MongoDB
-            Models.post.findByIdAndUpdate(req.params.id, req.body, (err, data) => {
-                // Check err
-                return err
-                ? reject(err)
-                : resolve(data)
+            // Get single post
+            Models.post.findById(req.params.id, (err, mongoPost) => {
+                // Check error
+                if( err ){ return reject(err) }
+                else{
+                    // Check author
+                    if( mongoPost.author === req.user._id ){
+                        // Get all post from MongoDB
+                        Models.post.findByIdAndUpdate(req.params.id, req.body, (err, data) => {
+                            // Check err
+                            return err
+                            ? reject(err)
+                            : resolve(data)
+                        })
+                    }
+                    else{ return reject('Unauthorized') };
+                }
             })
         })
     }
@@ -60,7 +94,7 @@ Functions
     const deleteOne = req => {
         return new Promise( (resolve, reject) => {
             // Get all post from MongoDB
-            Models.post.deleteOne({ _id: req.params.id }, (err, data) => {
+            Models.post.deleteOne({ _id: req.params.id, author: req.user._id }, (err, data) => {
                 // Check err
                 return err
                 ? reject(err)
